backend/requirements.txt
# Group 1: Core FastAPI and Pydantic
uvicorn
fastapi
pydantic
pydantic-settings
python-dotenv
email-validator

# Group 2: Database
pymongo>=4.0.0
motor>=3.0.0

# Group 3: Authentication
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt

# Group 4: Document processing
python-docx
pypdf
tiktoken

# Group 5: Testing and HTTP
pytest
pytest-asyncio
pytest-cov
httpx
pytest-mock
coverage
circuitbreaker
PyYAML>=6.0.1

# Group 6: AI and Vector DB (install last)
pinecone-client>=3.0.0
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-openai>=0.0.5
langchain-anthropic>=0.0.5
langchain-cohere>=0.0.5
langchain-text-splitters>=0.0.1
langgraph>=0.0.10

#
aiofiles